// src/theme/theme.types.ts
export type AppTheme = {
  id: string;
  name: string;
  title: string;

  logoUrl: string;
  faviconUrl: string;

  hideSubtitle: boolean;

  menuColor: string;

  primaryColor: string;
  secondaryColor: string;
  tertiaryColor: string;

  formItemFontSize: number;

  fontSizeButtonColor: string;
  fontSizeButtonFontColor: string;
};